moved show error logic into Input component
