+++ src/modules/common/corgi/actions/Transfer/Transfer.jsx
@@ -20,31 +20,11 @@ const Transfer = ({ id }) => {
 
   const [receiver, setReceiver] = useState('');
 
-  const [timeoutId, setTimeoutId] = useState(null);
-
   const [errorMessage, setErrorMessage] = useState('');
-  const [isErrorShown, setIsErrorShown] = useState(false);
-
-  const clearTimeoutId = () => {
-    if (timeoutId) {
-      clearTimeout(timeoutId);
-      setTimeoutId(null);
-    }
-  };
-
-  const showError = (error) => {
-    setErrorMessage(error);
-    setIsErrorShown(true);
-  };
-
-  const hideError = () => {
-    setIsErrorShown(false);
-    clearTimeoutId();
-  };
 
   const checkAccountLegit = async (newReceiver) => {
     if (newReceiver === user.accountId) {
-      showError(USER_VALIDATION_MESSAGES.OWN_ACCOUNT);
+      setErrorMessage(USER_VALIDATION_MESSAGES.OWN_ACCOUNT);
       return false;
     }
 
@@ -54,10 +34,10 @@ const Transfer = ({ id }) => {
       if (isAccountExist) {
         return true;
       } else {
-        showError(USER_VALIDATION_MESSAGES.NOT_EXIST);
+        setErrorMessage(USER_VALIDATION_MESSAGES.NOT_EXIST);
       }
     } catch (error) {
-      showError(USER_VALIDATION_MESSAGES.NOT_EXIST);
+      setErrorMessage(USER_VALIDATION_MESSAGES.NOT_EXIST);
 
       console.error(error);
     }
@@ -81,36 +61,6 @@ const Transfer = ({ id }) => {
     }
   };
 
-  useEffect(() => {
-    if (!timeoutId && isErrorShown) {
-      setTimeoutId(
-        setTimeout(() => {
-          hideError();
-        }, 5000),
-      );
-    }
-
-    return () => {
-      clearTimeoutId();
-    };
-  }, [timeoutId, isErrorShown, setTimeoutId, setIsErrorShown]);
-
-  useEffect(() => {
-    if (!timeoutId && !isErrorShown && errorMessage && errorMessage.length) {
-      setTimeoutId(
-        setTimeout(() => {
-          setErrorMessage('');
-
-          clearTimeoutId();
-        }, 1000),
-      );
-    }
-
-    return () => {
-      clearTimeoutId();
-    };
-  }, [timeoutId, isErrorShown, errorMessage, setTimeoutId, setErrorMessage]);
-
   return (
     <form className='transfer' onSubmit={(event) => onSubmit(event)}>
       <div className='transfer__input'>
@@ -120,7 +70,6 @@ const Transfer = ({ id }) => {
           placeholder='account.testnet'
           type='text'
           error={errorMessage}
-          showError={isErrorShown}
           required
         />
       </div>
